+++ CHANGELOG.md
@@ -8,11 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- Allow distinguishing tristate and Boolean encoding in KConfigReader-extracted formulas
+
 ### Changed
 
 ### Removed
 
 ### Fixed
+
+- Fixed bug concerning payload file extraction
 
 ## [2.1.0] - 2025-12-19